cygwin 1.7 build fixes
It's less accepting of signed char ... insisting that e.g. tolower()
not receive one as a parameter.

It's probably good to phase out such usage, given the number of bugs
that lurk in the vicinity (assumptions that char is unsigned), so fix
these even though such usage is actually legal.

+++ src/flash/mflash.c
@@ -1345,7 +1345,8 @@ COMMAND_HANDLER(mg_bank_cmd)
 	char *str;
 	mflash_bank->rst_pin.num = strtoul(CMD_ARGV[2], &str, 0);
 	if (*str)
-		mflash_bank->rst_pin.port[0] = (uint16_t)tolower(str[0]);
+		mflash_bank->rst_pin.port[0] = (uint16_t)
+				tolower((unsigned)str[0]);
 
 	mflash_bank->target = target;
